Checkin before we lose precision.

+++ timemachine/cpp/src/math_utils.cuh
@@ -1,6 +1,6 @@
 #pragma once
 
 template<typename T>
-__device__ __host__ int sign(T a) {
+__device__ __host__ double sign(T a) {
   return (a > 0) - (a < 0);
 }
